style(web-demo): move fillgroup components to folder
