+++ packages/pixelmanipulator-web-demo/src/components/fillgroup/FillButton.vue
